+++ react/components/ReviewBlock.tsx
@@ -1,5 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable vtex/prefer-early-return */
+/* eslint no-shadow: "error" */
 import React, {
   useState,
   FunctionComponent,
@@ -24,6 +25,7 @@ import { TARGET_SYSTEM } from '../utils/const'
 import { keyValuePairsToString } from '../utils/performanceDataProcessing'
 import ItemPricing from './ItemPricing'
 import StockAvailability from './StockAvailability'
+import ItemListContext from '../ItemListContext'
 import GET_PRODUCT_DATA from '../queries/getPrductAvailability.graphql'
 import './ReviewBlock.css'
 
@@ -33,14 +35,17 @@ const messages = defineMessages({
   valid: {
     id: 'store/quickorder.valid',
   },
-  prodnotfound: {
-    id: 'store/quickorder.productNotFound',
-  },
   available: {
     id: 'store/quickorder.available',
   },
   unavailable: {
     id: 'store/quickorder.unavailable',
+  },
+  unAuthorized: {
+    id: 'store/quickorder.unauthorized',
+  },
+  unAuthorizedError: {
+    id: 'store/quickorder.unauthorizedError',
   },
   invalidPattern: {
     id: 'store/quickorder.invalidPattern',
@@ -199,6 +204,8 @@ const CSS_HANDLES = [
   'orderedQuantity',
   'orderedQuantityLabel',
   'orderedQuantityValue',
+  'quickOrderTableRowWithIssues',
+  'unAuthorizedMessage',
 ]
 
 const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
@@ -206,7 +213,6 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
   reviewedItems,
   onRefidLoading,
   intl,
-  soldToAccount,
 }: any) => {
   // const { data: orderFormData } = useQuery<{
   //   orderForm
@@ -217,12 +223,16 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
   const client = useApolloClient()
   const styles = useCssHandles(CSS_HANDLES)
 
-  const customerNumber =
-    soldToAccount?.getOrderSoldToAccount?.customerNumber ?? ''
+  // const customerNumber =
+  //   soldToAccount?.getOrderSoldToAccount?.customerNumber ?? ''
 
-  const targetSystem = soldToAccount?.getOrderSoldToAccount?.targetSystem ?? ''
-  const salesOrganizationCode =
-    soldToAccount?.getOrderSoldToAccount?.salesOrganizationCode ?? ''
+  // const targetSystem = soldToAccount?.getOrderSoldToAccount?.targetSystem ?? ''
+  // const salesOrganizationCode =
+  //   soldToAccount?.getOrderSoldToAccount?.salesOrganizationCode ?? ''
+
+  const { useItemListState } = ItemListContext
+  const { customerNumber, targetSystem, salesOrganizationCode, itemStatuses } =
+    useItemListState()
 
   const [state, setReviewState] = useState<any>({
     reviewItems:
@@ -241,9 +251,10 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
   // }
   const errorMessage = {
     'store/quickorder.valid': messages.valid,
-    'store/quickorder.productNotFound': messages.prodnotfound,
     'store/quickorder.available': messages.available,
     'store/quickorder.unavailable': messages.unavailable,
+    'store/quickorder.unauthorized': messages.unAuthorized,
+    'store/quickorder.unauthorizedError': messages.unAuthorizedError,
     'store/quickorder.invalidPattern': messages.invalidPattern,
     'store/quickorder.skuNotFound': messages.skuNotFound,
     'store/quickorder.withoutStock': messages.withoutStock,
@@ -311,8 +322,8 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
         return item.sku !== null
       })
 
-      const refNotAvailable = itemsFromQuery.filter((item: any) => {
-        return item.availability !== 'available'
+      const refNotAuthorized = itemsFromQuery.filter((item: any) => {
+        return item.availability !== 'authorized'
       })
 
       const vtexSku = (item: any) => {
@@ -363,21 +374,6 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
         return ret
       }
 
-      // const getSellers = (item: any) => {
-      //   let ret: any = []
-      //
-      //   if (!!refidData && !!refidData.getSkuAvailability.items) {
-      //     ret = refidData.getSkuAvailability.items.find((curr: any) => {
-      //       return !!item.sku && item.sku === curr.refid
-      //     })
-      //     if (!!ret && !!ret.sellers) {
-      //       ret = ret.sellers
-      //     }
-      //   }
-      //
-      //   return ret
-      // }
-
       const errorMsg = (item: any) => {
         let ret: any = null
         const notfound = refIdNotFound.find((curr: any) => {
@@ -390,19 +386,18 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
 
         ret = notfound
           ? 'store/quickorder.skuNotFound'
-          : found?.availability && found.availability !== 'available'
-          ? `store/quickorder.${found.availability}`
-          : null
+          : found?.availability && found.availability === 'unauthorized'
+          ? 'store/quickorder.unauthorizedError'
+          : item.error
 
         return ret
       }
 
-      if (refIdNotFound.length || refNotAvailable.length) {
+      if (refIdNotFound.length || refNotAuthorized.length) {
         error = true
       }
 
       const items = reviewed.map((item: any) => {
-        // const sellers = getSellers(item)
         const sellers = item.sku ? mappedRefId[item.sku]?.sellers : '1'
         const itm = getItemFromQuery(item)
 
@@ -440,6 +435,7 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
       })
 
       onReviewItems(updated)
+
       setReviewState({
         ...state,
         reviewItems: updated,
@@ -545,6 +541,20 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
     })
   }
 
+  const getLineError = (lineItem: any) => {
+    const item = itemStatuses.find((itm: any) => itm.index === lineItem.index)
+
+    return item?.error ? intl.formatMessage(errorMessage[item.error]) : null
+  }
+
+  const getLineItemStatus = (lineItem: any) => {
+    const item = itemStatuses.find((itm: any) => itm.index === lineItem.index)
+
+    return item?.availability
+  }
+
+  const tableStyles = `.ReactVirtualized__Grid__innerScrollContainer>div {padding: 0;}`
+
   const tableSchema = {
     properties: {
       col1: {
@@ -554,16 +564,17 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
         }),
         // eslint-disable-next-line react/display-name
         cellRenderer: ({ rowData }: any) => {
-          const statusMessage = intl.formatMessage(
-            errorMessage[
-              rowData?.error !== null && rowData?.error !== undefined
-                ? rowData?.error
-                : 'store/quickorder.available'
-            ]
-          )
+          const itemError = getLineError(rowData)
+          const itemAvailability = getLineItemStatus(rowData)
 
           return (
-            <div className={`${styles.quickOrderTable} flex w-100 relative`}>
+            <div
+              className={`${styles.quickOrderTable} ${
+                itemError ? styles.quickOrderTableRowWithIssues : ''
+              } flex w-100 relative`}
+            >
+              <style>{tableStyles}</style>
+
               <div
                 className={`${styles.tableCol1} flex flex-column ${
                   targetSystem === TARGET_SYSTEM.JDE ? 'w-40' : 'w-50'
@@ -599,28 +610,31 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
                       Ordered Quantity:
                     </span>
                     <span className={`${styles.orderedQuantityValue}`}>
-                      {Number.isNaN(rowData.quantity)
-                        ? intl.formatMessage(messages.prodnotfound)
-                        : rowData.quantity}
+                      {!Number.isNaN(rowData.quantity) ? rowData.quantity : ''}
                     </span>
                   </div>
                   <div className={`${styles.productPageLink} ml3`}>
-                    <a
-                      className={`${styles.productDetailsLink} flex-column`}
-                      href={`${rowData?.linkText}/p`}
-                      target="_blank"
-                      rel="noreferrer"
-                    >
-                      {Number.isNaN(rowData.quantity)
-                        ? ''
-                        : intl.formatMessage(messages.goToProductPage)}
-                    </a>
+                    {!Number.isNaN(rowData.quantity) && rowData?.linkText && (
+                      <a
+                        className={`${styles.productDetailsLink} flex-column`}
+                        href={`${rowData?.linkText}/p`}
+                        target="_blank"
+                        rel="noreferrer"
+                      >
+                        {intl.formatMessage(messages.goToProductPage)}
+                      </a>
+                    )}
                   </div>
+                </div>
+                <div>
+                  <span className="red">{itemError}</span>
                 </div>
               </div>
 
               <div
-                className={`${styles.tableCol2} flex pa3 ${
+                className={`${
+                  itemError ? styles.quickOrderTableRowWithIssues : ''
+                } ${styles.tableCol2} flex pa3 ${
                   targetSystem === TARGET_SYSTEM.JDE ? 'w-40' : 'w-30'
                 }`}
               >
@@ -685,23 +699,26 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
                 <div
                   className={`${styles.stockAvailabilityMessage} flex w-100`}
                 >
-                  <Tooltip label={statusMessage}>
-                    {rowData?.availability === 'available' ||
-                    statusMessage === null ||
-                    statusMessage === '' ? (
-                      <span className={`${styles.inStockMessage} b ttu`}>
-                        {intl.formatMessage(messages.inStock)}
+                  {itemAvailability === 'available' ? (
+                    <span className={`${styles.inStockMessage} b ttu`}>
+                      {intl.formatMessage(messages.inStock)}
+                    </span>
+                  ) : itemAvailability === 'unavailable' && !itemError ? (
+                    <span className={`${styles.outOfStockMessage} b ttu`}>
+                      {intl.formatMessage(messages.outOfStock)}
+                    </span>
+                  ) : (
+                    itemAvailability === 'unauthorized' && (
+                      <span className={`${styles.unAuthorizedMessage} b ttu`}>
+                        {intl.formatMessage(messages.unAuthorized)}
                       </span>
-                    ) : (
-                      <span className={`${styles.outOfStockMessage} b ttu`}>
-                        {intl.formatMessage(messages.outOfStock)}
-                      </span>
-                    )}
-                  </Tooltip>
+                    )
+                  )}
                 </div>
                 <div className={`${styles.stockAvailabilityValue} w-40`}>
                   {targetSystem === TARGET_SYSTEM.JDE ? (
                     <StockAvailability
+                      itemIndex={rowData?.index}
                       itemNumber={rowData?.sku}
                       customerNumber={customerNumber}
                     />